requirements.txt
langchain
langchain-google-genai
langchain-community
ollama
google-genai